+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -8,7 +8,7 @@ import nextstep.jwp.exception.UncheckedServletException;
 import org.apache.coyote.Processor;
 import org.apache.coyote.http.request.HttpRequest;
 import org.apache.coyote.http.response.HttpResponse;
-import org.apache.coyote.http.RequestMapping;
+import org.apache.coyote.controller.RequestMapping;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
