+++ tomcat/src/main/java/org/apache/coyote/session/SessionManager.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.session;
 
 import java.util.HashMap;
 import java.util.Map;
